backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.23.2
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
redis==5.0.1
kafka-python==2.0.2
PyJWT==2.8.0
pydantic==1.10.13
